Update waist measurement to allow decimal values
Modify the database schema and frontend parsing to support decimal values for waist measurements, replacing integer and parseInt with real and parseFloat respectively.

+++ client/src/pages/profile-page.tsx
@@ -388,7 +388,7 @@ export default function ProfilePage({ onClose }: ProfilePageProps) {
         ...(data.weight &&
           data.weight !== "" && { weight: parseInt(String(data.weight)) }),
         ...(data.waist &&
-          data.waist !== "" && { waist: parseInt(String(data.waist)) }),
+          data.waist !== "" && { waist: parseFloat(String(data.waist)) }),
       };
 
       console.log("Submitting measurement:", payload);
